move model only tests to a new file
