+++ qt/python/mantidqtinterfaces/test/Muon/phase_table_widget/phase_table_presenter_test.py
@@ -67,25 +67,6 @@ class PhaseTablePresenterTest(unittest.TestCase):
 
         self.assertEqual(self.model.phase_context.options_dict["input_workspace"], workspace_name)
 
-    def test_create_parameters_for_cal_muon_phase_returns_correct_parameter_dict(self):
-        # TODO: Move to new model test
-        workspace_name = "input_workspace_name_raw_data"
-        self.model.phase_context.options_dict["input_workspace"] = workspace_name
-
-        result = self.model.create_parameters_for_cal_muon_phase_algorithm()
-
-        self.assertEqual(
-            result,
-            {
-                "BackwardSpectra": [2, 4, 6, 8, 10],
-                "FirstGoodData": 0.1,
-                "ForwardSpectra": [1, 3, 5, 7, 9],
-                "InputWorkspace": workspace_name,
-                "LastGoodData": 15,
-                "DetectorTable": "input_workspace_name; PhaseTable; fwd; bwd",
-            },
-        )
-
     def test_correctly_retrieves_workspace_names_associsated_to_current_runs(self):
         self.view.set_input_combo_box = mock.MagicMock()
         self.model.get_grouped_workspace_names = mock.MagicMock(return_value=["MUSR22222", "MUSR44444"])
@@ -100,17 +81,6 @@ class PhaseTablePresenterTest(unittest.TestCase):
         self.presenter.update_current_groups_list()
 
         self.view.set_group_combo_boxes.assert_called_once_with(["fwd", "bwd"])
-
-    @mock.patch("mantidqtinterfaces.Muon.GUI.Common.phase_table_widget.phase_table_model.MuonWorkspaceWrapper")
-    def test_that_phase_table_added_to_ADS_with_correct_name_and_group(self, mock_workspace_wrapper):
-        # TODO: Move to new model test
-        workspace_wrapper = mock.MagicMock()
-        mock_workspace_wrapper.return_value = workspace_wrapper
-
-        self.model.add_phase_table_to_ads("MUSR22222_period_1; PhaseTable")
-
-        mock_workspace_wrapper.assert_called_once_with("MUSR22222 MA/MUSR22222_period_1; PhaseTable")
-        workspace_wrapper.show.assert_called_once_with()
 
     @mock.patch("mantidqtinterfaces.Muon.GUI.Common.phase_table_widget.phase_table_model.run_CalMuonDetectorPhases")
     def test_handle_calculate_phase_table_clicked_behaves_correctly_for_succesful_calculation(self, run_algorithm_mock):
